** fixed compile errors
git-svn-id: https://open-ms.svn.sourceforge.net/svnroot/open-ms/branches/Release1.6@6247 6adb6e08-d915-0410-941f-83917bcadc18

+++ source/FORMAT/Bzip2Ifstream.C
@@ -64,7 +64,7 @@ namespace OpenMS
 		if(bzip2file != NULL)
 		{
 			bzerror = BZ_OK;
-  		n_buffer = BZ2_bzRead ( &bzerror, bzip2file, s, n/* size of buf */ );		
+  		n_buffer = BZ2_bzRead ( &bzerror, bzip2file, s, (unsigned int)n/* size of buf */ );		
 	  	if(bzerror == BZ_OK) 
 	 		{
     			return n_buffer;
